refactor gradient example plugin into bev plugin
