+++ boat2023_nav2/params/carla_nav2_params.yaml
@@ -149,9 +149,9 @@ local_costmap:
       resolution: 0.05
       robot_radius: 0.2
       always_send_full_costmap: true
-      plugins: ["gradient_layer", "inflation_layer"]
-      gradient_layer:
-        plugin: "nav2_gradient_costmap_plugin/GradientLayer"
+      plugins: ["bev_layer", "inflation_layer"]
+      bev_layer:
+        plugin: "nav2_bev_costmap_plugin/BEVLayer"
         enabled: true
       inflation_layer:
         plugin: "nav2_costmap_2d::InflationLayer"
@@ -175,9 +175,9 @@ global_costmap:
       resolution: 0.05
       robot_radius: 0.2
       always_send_full_costmap: true
-      plugins: ["gradient_layer", "inflation_layer"]
-      gradient_layer:
-        plugin: "nav2_gradient_costmap_plugin/GradientLayer"
+      plugins: ["bev_layer", "inflation_layer"]
+      bev_layer:
+        plugin: "nav2_bev_costmap_plugin/BEVLayer"
         enabled: true
       inflation_layer:
         plugin: "nav2_costmap_2d::InflationLayer"
