+++ nav2_bev_costmap_plugin/include/nav2_bev_costmap_plugin/bev_layer.hpp
@@ -1,5 +1,5 @@
-#ifndef GRADIENT_LAYER_HPP_
-#define GRADIENT_LAYER_HPP_
+#ifndef BEV_LAYER_HPP_
+#define BEV_LAYER_HPP_
 
 #include "rclcpp/rclcpp.hpp"
 #include "nav2_costmap_2d/layer.hpp"
@@ -8,13 +8,13 @@
 #include <opencv2/opencv.hpp>
 #include "sensor_msgs/msg/image.hpp"
 
-namespace nav2_gradient_costmap_plugin
+namespace nav2_bev_costmap_plugin
 {
 
-class GradientLayer : public nav2_costmap_2d::Layer
+class BEVLayer : public nav2_costmap_2d::Layer
 {
 public:
-  GradientLayer();
+  BEVLayer();
 
   virtual void onInitialize();
   virtual void updateBounds(
@@ -43,7 +43,7 @@ private:
 
   double last_min_x_, last_min_y_, last_max_x_, last_max_y_;
 
-  // Indicates that the entire gradient should be recalculated next time.
+  // Indicates that the entire BEV gradient should be recalculated next time.
   bool need_recalculation_;
 
   // Size of gradient in cells
@@ -52,6 +52,6 @@ private:
   int GRADIENT_FACTOR = 10;
 };
 
-}  // namespace nav2_gradient_costmap_plugin
+}  // namespace nav2_bev_costmap_plugin
 
-#endif  // GRADIENT_LAYER_HPP_
+#endif  // BEV_LAYER_HPP_
