--- nav2_gradient_costmap_plugin/gradient_layer.xml
@@ -1,7 +1,0 @@
-<class_libraries>
-  <library path="nav2_gradient_costmap_plugin">
-    <class name="nav2_gradient_costmap_plugin/GradientLayer" type="nav2_gradient_costmap_plugin::GradientLayer" base_class_type="nav2_costmap_2d::Layer">
-      <description>This is an example plugin which puts repeating costs gradients to costmap</description>
-    </class>
-  </library>
-</class_libraries>
